add sparse flag to toggle sizeof(void) warnings
Recent changes to the min()/max() macros in include/linux/kernel.h
have added a lot of noise to sparse output; this is due to the
*huge* number of new sizeof(void) warnings, most of which can
largely be ignored.

Add the -Wsizeof_void flag to enable/disable these warnings on
demand; the warning itself has been disabled by default to reduce the
large influx of noise which was inadvertently added by commit
3c8ba0d61d04ced9f8 (kernel.h: Retain constant expression output
for max()/min()).

Update the manpage to document the new flag.

CC: Kees Cook <keescook@chromium.org>
CC: Linus Torvalds <torvalds@linux-foundation.org>
CC: Martin Uecker <Martin.Uecker@med.uni-goettingen.de>
CC: Al Viro <viro@ZenIV.linux.org.uk>
CC: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
CC: Christopher Li <sparse@chrisli.org>
Signed-off-by: Joey Pabalinas <joeypabalinas@gmail.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>